fix(repository): fix file name same as module name by moving repository into mod.rs
